Hotfix: Fix `pretty_run_time` displaying minutes and seconds as floats
This fixes the `pretty_run_time` from displaying minutes and
seconds as floats instead of integers. This is confusing because
it doesn't fit the formatting nicely and the minutes shows
duplicate information (if it took 1 min and 30 seconds it would
display 1.5:30.0).

This also changes the amount of time where it chooses which format
to show. This was switched to 1.3 from 1.0 to keep showing the
detail when the timing is close to a second. This splits the
difference between 1.5 and 1.0 to get 1.3.

Finally, this fixes a broken test from 683d796. An extra log line
was added when logging task runs but the tests weren't updated to
expect the new log line.