update non-verbose logs to no longer show task starts/ends
This is a UX effort to keep the default log output as clean and
readable as possible.  This removes the "chunked" newlines before
and after task start/end and switches individual task run start
and end logs to be at the debug level.  This gets them off by
default.  Then, so that we don't have rando cmds/logs with no
context, this updates each top level cli task run to log at the
info level its start end.

I like this better b/c it is more clean and highlights the cmds
being run and any custom log more than what tasks are being run.
The start/stop logs that are displayed correspond to the user
input which makes it intuitive.

```
$ dk my-task my-other-task
Starting `my-task`.
      running the sub task
      running what the sub task does
[CMD] <some cmd str>
      (200.0ms)
      > cmd output line 1
      > cmd output line 2
[SSH] <some cmd str>
      ssh -o ...  <host> -- "sh -c \"<some cmd str>\""
      [host1]
      [host2]
      (913.5ms)
`my-task` finished in 0:01s.

Starting `my-other-task`
      doing what the other task does
`my-other-task` finished in 347.3ms.

(0:01s)
$
```

Note: this also tweaks any reference to a task name to wrap it
in backticks - this looks "more official" than double-quotes.