(hotfix) have remote cmds expose their stdout and stderr
This updates the Remote base cmd to expose these attrs.  It looks
through the local cmds and just chooses the first one that isn't
empty for each.  This way if only one of the hosts errors, you'll
still get stderr info even if it wasn't the first in the list.

The whole point of this is so that you can make ssh calls against
a single host and use the call's output.  This allows you to use
all the features of ssh calls and not have to manually build ssh
local cmds just to get their output.