update stubs API to optionally take proc values
This allows you to specify the cmd str, input and cmd opts like
the task implementation does.  This makes stubbing cmds that are
driven by dynamic task instance values possible.  If the stubs
are given procs, those procs are instance eval'd against the task
when looking up a matching stub for a cmd.

A side effect of this change is we have to alter the API of the
config/runner stub methods.  The input and opts are now specified
with an hash instead of fuzzy args.  This is b/c we can't evaluate
the fuzzy args and understand them if they are procs.  We can't
tell if a proc is for the input or the cmd opts.