add a custom exception for notifying the user of problems
This new `Dk::NoticeError` can be raised in tasks for notifying
the user of a problem.  This stops execution and displays the
exception message on its own followed by the first line of the
backtrace.  The goal is to provide a way to stop execution while
communicating a message to the user (without the full backtrace
causing the message to flow off terminal).

For example:

```
$ dk deploy -d
Starting `deploy`.
      Downloading servers list...
[SSH] cat servers.yaml
      [host1]
      (308.0ms)
      Checking that the `deploy:setup` task has been run...
[SSH] test -d /deploy-dir1 && test -d /deploy-dir1/.git
      [host1]
      [host2]
      (0.1ms)

**** some of the servers are not setup for deploying ****
****       run the `deploy:setup` task first         ****

/path/to/my-dk-deploy/lib/my-dk-deploy/validate_setup.rb:33:in `run!'
```