(hotfix) switch to "dk" name for default task config file
This is more explicit about what it's a config for.  This also
updates the log file example in the README to use `log/dk.log`.
In using Dk in some apps and 3rd-party libs, this has proven to
be a better default.  Plus there is less chance of name collisions
with another config file.

Collin and I discussed this and beat it up before coming to this
decision.