(hotfix) tweak time displays to cross over at 1.5s
This is just a minor tweak to the pretty time displays.  I decided
1.5 was a better cross over mark than 1.3.  See 0:01 is not so
helpful in the tests I've run.  If the time is less than 1.5s,
this now always shows the time in milliseconds.  If it is greater
than 1.5s, it shows it in seconds.  This will allow more times
that are really close to 1 second to be shown in milliseconds
which is more helpful and revealing.

This also changes the seconds display to round to the nearest
second.  This provides more accurate times.  The side effect of
this change and the change above is that now you will never see
`0:01s`.  If it is less that 1.5s, you will see milliseconds; if
it is more than 1.5s you will see `0:02s`.  Overall this side
effect is not important compared to having more accurate and
meaningful time displays.