add a `start` helper for asynchronously running local system commands
This complements the `cmd` and `cmd!` helpers. Like those, this
runs a local system command but it doesn't block waiting on an
exitstatus.

This is handy for running system commands that run while you also
run additional system commands. The classic example would be
running a server, then running other system commands that interact
with that server, and finally running a command to signal that
server to shutdown. In this example you would use `start` to run
the local server and use `cmd`/`cmd!` to interact with and shutdown
the server.

This also includes all the support logic for this new helper to
make it work just like `cmd`/`cmd!`: spy logic, test runner, etc.